+++ pallets/relayer/src/benchmarking.rs
@@ -71,13 +71,13 @@ pub fn add_non_syncing_validators<T: Config>(
 
 benchmarks! {
   register {
-    let p in 0 .. T::MaxProgramHashes::get();
+    let p in 1 .. T::MaxProgramHashes::get();
     let program = vec![0u8];
     let program_hash = T::Hashing::hash(&program);
     let program_hashes = vec![program_hash; p as usize];
     let bounded_program_hashes = BoundedVec::try_from(program_hashes).unwrap();
     let program_modification_account: T::AccountId = whitelisted_caller();
-    Programs::<T>::insert(program_hash, ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone()});
+    Programs::<T>::insert(program_hash, ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone(), ref_counter: 0});
     let sig_req_account: T::AccountId = whitelisted_caller();
     let balance = <T as pallet_staking_extension::Config>::Currency::minimum_balance() * 100u32.into();
     let _ = <T as pallet_staking_extension::Config>::Currency::make_free_balance_be(&sig_req_account, balance);
@@ -88,10 +88,12 @@ benchmarks! {
   }
 
   prune_registration {
+    let p in 1 .. T::MaxProgramHashes::get();
     let program_modification_account: T::AccountId = whitelisted_caller();
     let program = vec![0u8];
     let program_hash = T::Hashing::hash(&program);
     let program_hashes = BoundedVec::try_from(vec![program_hash]).unwrap();
+    Programs::<T>::insert(program_hash, ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone(), ref_counter: 1});
     let sig_req_account: T::AccountId = whitelisted_caller();
     let balance = <T as pallet_staking_extension::Config>::Currency::minimum_balance() * 100u32.into();
     let _ = <T as pallet_staking_extension::Config>::Currency::make_free_balance_be(&sig_req_account, balance);
@@ -108,24 +110,27 @@ benchmarks! {
   }
 
   change_program_pointer {
-    let p in 0 .. T::MaxProgramHashes::get();
+    let n in 1 .. T::MaxProgramHashes::get();
+    let o in 1 .. T::MaxProgramHashes::get();
     let program_modification_account: T::AccountId = whitelisted_caller();
     let program = vec![0u8];
     let program_hash = T::Hashing::hash(&program);
-    let program_hashes = BoundedVec::try_from(vec![program_hash]).unwrap();
+    let program_hashes = vec![program_hash; o as usize];
+    let bounded_program_hashes = BoundedVec::try_from(vec![program_hash]).unwrap();
     let new_program = vec![1u8];
     let new_program_hash = T::Hashing::hash(&new_program);
-    let new_program_hashes = vec![new_program_hash; p as usize];
+    let new_program_hashes = vec![new_program_hash; n as usize];
     let new_bounded_program_hashes = BoundedVec::try_from(new_program_hashes).unwrap();
     let sig_req_account: T::AccountId = whitelisted_caller();
-    Programs::<T>::insert(new_program_hash, ProgramInfo {bytecode: new_program, program_modification_account: program_modification_account.clone()});
+    Programs::<T>::insert(program_hash, ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone(), ref_counter: 0});
+    Programs::<T>::insert(new_program_hash, ProgramInfo {bytecode: new_program, program_modification_account: program_modification_account.clone(), ref_counter: o as u128});
     let balance = <T as pallet_staking_extension::Config>::Currency::minimum_balance() * 100u32.into();
     let _ = <T as pallet_staking_extension::Config>::Currency::make_free_balance_be(&sig_req_account, balance);
     <Registered<T>>::insert(
         &sig_req_account,
         RegisteredInfo {
             program_modification_account: sig_req_account.clone(),
-            program_pointers: program_hashes,
+            program_pointers: bounded_program_hashes,
             verifying_key: BoundedVec::default(),
             key_visibility: KeyVisibility::Public,
         },