+++ CHANGELOG.md
@@ -28,6 +28,7 @@ At the moment this project **does not** adhere to
 - Add password file option ([#555](https://github.com/entropyxyz/entropy-core/pull/555))
 - Include contents of x25515chacha20poly1305 repo in entropy-protocol ([#563](https://github.com/entropyxyz/entropy-core/pull/563))
 - Custom Hashing Algorithms [#553](https://github.com/entropyxyz/entropy-core/pull/553/)
+- Add ref counter to programs [#585](https://github.com/entropyxyz/entropy-core/pull/585/)
 
 ### Changed
 - Crate name refactor ([#561](https://github.com/entropyxyz/entropy-core/pull/561))